Add taskrunner for > 1k simultaneous jobs
Our UGER cluster limits 1k simultaneous uger jobs (pending + running) at any
time for each user. For large S4 runs we were easily going over that limit. So
instead of directly `call`'ing the qsub commands, this adds a layer of
abstraction with `call_to_taskrunner`. The jobs submit the new qsub jobs as a
separate fike to a temporary tmp_taskrunner folder (with especial care [1] to
avoid filename collisions). Then the new_taskrunner job reads these files,
checks if there is < 1k headroom and runs the qsub job, moving the file to a
"done" directory. (In case the user is running non-traskrun jobs, to avoid a
thundering herd situation, it tries to ease up to the max umber of jobs)

[1] It uses datetime up to the microsecond in the filename and adds an extra set
of random characters for perfectly simulataneous writes